services/price_predictor/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["ameyshinde28 <aammeeyy28@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
hopsworks = "^3.7.0"
pydantic-settings = "^2.5.2"
loguru = "^0.7.2"
scikit-learn = "^1.5.2"
comet-ml = "^3.47.0"
xgboost = "^2.1.1"
ta-lib = "^0.4.32"
optuna = "^4.0.0"
fastapi = "^0.115.5"
uvicorn = "^0.32.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
